Tasks

+++ Tasks/3-V/3-V.cpp
@@ -2,7 +2,7 @@
 #include <Windows.h>
 using namespace std;
 
-double function(int x)
+double function(double x)
 {
 	if (pow(x, 2) - 2 == 0 || pow(x, 3) - 1 < 0) return 'E';
 	return (x + 4) / (pow(x, 2) - 2) + sqrt(pow(x, 3) - 1);
@@ -12,7 +12,7 @@ int main()
 {
 	SetConsoleCP(1251);
 	SetConsoleOutputCP(1251);
-	int a, b, h;
+	double a, b, h;
 	cout << "Введите значение A, B и H через пробел\n";
 	cin >> a >> b >> h;
 	cout << "Значение X\tЗначение функции\n";
